requirements.txt
Flask==2.3.2
flask_sqlalchemy==3.0.5
google_api_python_client==2.93.0
google_auth_oauthlib==1.0.0
pandas==2.0.3
protobuf==4.23.4
SQLAlchemy==2.0.18
